Added progress bar tracking function.
Progress bar tracking is hard coded, looking for "challenge" and "trial" to update. lib\constants.ts file looks hard codes how many trials or challenges are in each lesson. Tried to do this dynamically with a function but didn't work. Maybe look into this in the future, but for now leave it hard coded.